teams/tim-03/backend/requirements.txt
fastapi[standard]>=0.128.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pillow>=10.0.0
uvicorn[standard]>=0.27.0
